RFC: Mirrors callback structure
