+++ skorch/tests/callbacks/test_lr_scheduler.py
