+++ skorch/tests/callbacks/test_regularization.py
