+++ skorch/tests/callbacks/test_training.py
